With 64 bits indices PetscInt is now long int. Therefore, the functions expecting (int) output a warning when a PetscInt is passed. To remove those warnings (int) has been add in front of the variables when they are called by functions expecting int and not long int (as printf() for exemple). This cast is safe as long as the integer is not larger than 2^32. Moreover, when DataBucketGetSizes() is called, the variable representing the number of points should be declared as an int and not a long int. Therefore it is safer to declare an int and not a PetscInt.

+++ src/finite_volume/fv-ex3-fas.c
@@ -299,7 +299,7 @@ PetscErrorCode t3_warp_mms(void)
     m[1] *= 2;
     m[2] *= 2;
     ierr = create_level(m,&fvl[k]);CHKERRQ(ierr);
-    printf("  fvl[%d] %p\n",k,fvl[k]);
+    printf("  fvl[%d] %p\n",(int)k,fvl[k]);
   }
   
   fv = fvl[nlevels-1];
@@ -307,7 +307,7 @@ PetscErrorCode t3_warp_mms(void)
   for (k=0; k<nlevels; k++) {
     ierr = DMClone(fvl[k]->dm_fv,&dmc[k]);CHKERRQ(ierr);
     ierr = DMDASetInterpolationType(dmc[k],DMDA_Q0);CHKERRQ(ierr);
-    printf("dmc[%d] %p\n",k,dmc[k]);
+    printf("dmc[%d] %p\n",(int)k,dmc[k]);
   }
 
   /*
@@ -432,7 +432,7 @@ PetscErrorCode t3_warp_mms_r(void)
     m[2] *= rfac;
     ierr = create_level(m,&fvl[k]);CHKERRQ(ierr);
     ierr = DMDASetRefinementFactor(fvl[k]->dm_fv,rfac,rfac,rfac);CHKERRQ(ierr);
-    printf("  fvl[%d] %p\n",k,fvl[k]);
+    printf("  fvl[%d] %p\n",(int)k,fvl[k]);
   }
   
   fv = fvl[nlevels-1];
@@ -453,7 +453,7 @@ PetscErrorCode t3_warp_mms_r(void)
   printf("dmc[0] %p\n",dmc[0]);
   for (k=1; k<nlevels; k++) {
     ierr = DMRefine(dmc[k-1],PETSC_COMM_WORLD,&dmc[k]);CHKERRQ(ierr);
-    printf("dmc[%d] %p\n",k,dmc[k]);
+    printf("dmc[%d] %p\n",(int)k,dmc[k]);
     ierr = DMDASetInterpolationType(dmc[k],DMDA_Q0);CHKERRQ(ierr);
     ierr = DMDASetRefinementFactor(dmc[k],rfac,rfac,rfac);CHKERRQ(ierr);
   }
